# gatewayservice/openapi.yaml
openapi: 3.0.0
info:
  title: Gatewayservice API
  description: Gateway OpenAPI specification.
  version: 0.2.0
servers:
  - url: http://localhost:8000
    description: Development server
  - url: http://SOMEIP:8000
    description: Production server
paths:
...
  /api/game/endAndSaveGame:
    post:
      summary: Save the result of a completed game
      operationId: endAndSaveGame
      security:
        - bearerAuth: []  # Requiere autenticación JWT
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - token
                - numberOfQuestions
                - numberOfCorrectAnswers
                - gameMode
                - points
                - questions
              properties:
                numberOfQuestions:
                  type: integer
                  example: 30
                numberOfCorrectAnswers:
                  type: integer
                  example: 18
                gameMode:
                  type: string
                  example: normal
                points:
                  type: integer
                  example: 500
                questions:
                  type: array
                  items:
                    type: object
                    required:
                      - text
                      - imageUrl
                      - selectedAnswer
                      - answers
                    properties:
                      text:
                        type: string
                        example: Esto es una pregunta
                      imageUrl:
                        type: string
                        format: uri
                        example: https://example.com/question-image.jpg
                      selectedAnswer:
                        type: string
                        example: Esto es una respuesta1
                      answers:
                        type: array
                        items:
                          type: object
                          required:
                            - text
                            - isCorrect
                          properties:
                            text:
                              type: string
                              example: Esto es una respuesta1
                            isCorrect:
                              type: boolean
                              example: true
                        example:
                          - text: Esto es una respuesta1
                            isCorrect: true
                          - text: Esto es una respuesta2
                            isCorrect: false
                          - text: Esto es una respuesta3
                            isCorrect: false
                          - text: Esto es una respuesta4
                            isCorrect: false
      responses:
        '200':
          description: Game saved successfully.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Game data saved successfully.
        '400':
          description: Invalid input or missing fields.
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: Missing fields or invalid format
        '401':
          description: Unauthorized - Invalid or missing token.
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: Unauthorized
        '500':
          description: Error saving the game.
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: Internal server error
  /api/game/history/gameList:
    get:
      summary: Get game history for the authenticated user
      description: Returns a list of stored games for the currently authenticated user.
      operationId: getGameHistory
      security:
        - bearerAuth: []  # Requiere autenticación JWT
      responses:
        '200':
          description: List of games retrieved successfully.
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    _id:
                      type: string
                      example: 661c8e10f7c1a5c7b0b3d2a4
                    numberOfQuestions:
                      type: integer
                      example: 10
                    numberOfCorrectAnswers:
                      type: integer
                      example: 8
                    gameMode:
                      type: string
                      example: normal
                    points:
                      type: integer
                      example: 800
                    topics:
                      type: array
                      items:
                        type: string
                      example: ["art", "history"]
        '401':
          description: Unauthorized - Invalid or missing token.
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: Unauthorized
        '500':
          description: Internal server error.
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: Internal Server Error
        '404':
          description: No games found for the user.
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: No games found for this user.
  /api/game/new:
    post:
      summary: Configure a new game with selected topics and language
      operationId: configureNewGame
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - topics
                - lang
              properties:
                topics:
                  type: array
                  items:
                    type: string
                  example: ["art", "history"]
                lang:
                  type: string
                  enum: [es, en]
                  example: es
      responses:
        '200':
          description: Game configured successfully. Returns a cacheId to be used in next requests.
          content:
            application/json:
              schema:
                type: object
                properties:
                  cacheId:
                    type: string
                    example: abc123def456ghi789
        '400':
          description: Bad request - Invalid topics or language.
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: Invalid topics or language
        '500':
          description: Internal server error.
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: Internal Server Error
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT